+++ backend/src/AnimalAllies.API/Controllers/VolunteerController.cs
@@ -3,6 +3,8 @@ using AnimalAllies.API.Response;
 using AnimalAllies.Application.Contracts.DTOs;
 using AnimalAllies.Application.Features.Volunteer;
 using AnimalAllies.Application.Features.Volunteer.Create;
+using AnimalAllies.Application.Features.Volunteer.CreateRequisites;
+using AnimalAllies.Application.Features.Volunteer.CreateSocialNetworks;
 using AnimalAllies.Application.Features.Volunteer.Update;
 using AnimalAllies.Domain.Models;
 using FluentValidation;
@@ -59,7 +61,7 @@ public class VolunteerController: ApplicationController
     public async Task<IActionResult> CreateRequisitesToVolunteer(
         [FromRoute] Guid id,
         [FromBody] RequisiteListDto dto,
-        [FromServices] CreateRequisitesToVolunteerHandler handler,
+        [FromServices] CreateRequisitesHandler handler,
         [FromServices] IValidator<CreateRequisitesRequest> validator,
         CancellationToken cancellationToken = default)
     {
@@ -86,7 +88,7 @@ public class VolunteerController: ApplicationController
     public async Task<IActionResult> CreateSocialNetworksToVolunteer(
         [FromRoute] Guid id,
         [FromBody] SocialNetworkListDto dto,
-        [FromServices] CreateSocialNetworksToVolunteerHandler handler,
+        [FromServices] CreateSocialNetworksHandler handler,
         [FromServices] IValidator<CreateSocialNetworksRequest> validator,
         CancellationToken cancellationToken = default)
     {
