+++ backend/src/AnimalAllies.Application/Features/Volunteer/CreateRequisites/CreateRequisitesRequest.cs
@@ -1,5 +1,5 @@
 using AnimalAllies.Application.Contracts.DTOs;
 
-namespace AnimalAllies.Application.Features.Volunteer.Create;
+namespace AnimalAllies.Application.Features.Volunteer.CreateRequisites;
 
 public record CreateRequisitesRequest(Guid Id,RequisiteListDto Dto);
