+++ backend/src/AnimalAllies.Application/Features/Volunteer/CreateSocialNetworks/CreateSocialNetworksRequest.cs
@@ -1,5 +1,5 @@
 using AnimalAllies.Application.Contracts.DTOs;
 
-namespace AnimalAllies.Application.Features.Volunteer.Create;
+namespace AnimalAllies.Application.Features.Volunteer.CreateSocialNetworks;
 
 public record CreateSocialNetworksRequest(Guid Id,SocialNetworkListDto Dto);
