+++ backend/src/AnimalAllies.Application/Repositories/IVolunteerRepository.cs
@@ -9,7 +9,7 @@ public interface IVolunteerRepository
 {
     Task<Result<VolunteerId>> Create(Volunteer entity, CancellationToken cancellationToken = default);
     Task Delete(Volunteer entity, CancellationToken cancellationToken = default);
-    Task<Result<VolunteerId>> Update(Volunteer entity, CancellationToken cancellationToken = default);
+    Task<Result<VolunteerId>> Save(Volunteer entity, CancellationToken cancellationToken = default);
     Task<Result<Volunteer>> GetById(VolunteerId id, CancellationToken cancellationToken = default);
     Task<Result<Volunteer>> GetByPhoneNumber(PhoneNumber phone, CancellationToken cancellationToken = default);
     Task<Result<Volunteer>> GetByEmail(Email email, CancellationToken cancellationToken = default);
